+++ tests/testthat/test_spwb.R
@@ -209,7 +209,7 @@ test_that("spwb_day gives same result with inner and direct calls with general c
 
 })
 
-test_that("spwb_day does not produce warnings with truncated root distirbution",{
+test_that("spwb_day can be run with truncated root distribution",{
   exampleforest$shrubData$Z50 <- 100
   exampleforest$shrubData$Z95 <- 200
   control_granier$truncateRootDistribution<- TRUE
@@ -221,6 +221,26 @@ test_that("spwb_day does not produce warnings with truncated root distirbution",
   expect_s3_class(spwb_day(x2, date, meteovec,
                            latitude = 41.82592, elevation = 100, slope=0, aspect=0), "spwb_day")
   control_sureau$truncateRootDistribution<- TRUE
+  x3 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_sureau)
+  expect_s3_class(spwb_day(x3, date, meteovec,
+                           latitude = 41.82592, elevation = 100, slope=0, aspect=0), "spwb_day")
+})
+
+test_that("spwb_day can be run with truncated root distribution and rhizosphere overlap",{
+  exampleforest$shrubData$Z50 <- 100
+  exampleforest$shrubData$Z95 <- 200
+  control_granier$truncateRootDistribution<- TRUE
+  control_granier$rhizosphereOverlap <- "partial"
+  x1 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_granier)
+  expect_s3_class(spwb_day(x1, date, meteovec,
+                           latitude = 41.82592, elevation = 100, slope=0, aspect=0), "spwb_day")
+  control_sperry$truncateRootDistribution<- TRUE
+  control_sperry$rhizosphereOverlap <- "partial"
+  x2 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_sperry)
+  expect_s3_class(spwb_day(x2, date, meteovec,
+                           latitude = 41.82592, elevation = 100, slope=0, aspect=0), "spwb_day")
+  control_sureau$truncateRootDistribution<- TRUE
+  control_sureau$rhizosphereOverlap <- "partial"
   x3 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_sureau)
   expect_s3_class(spwb_day(x3, date, meteovec,
                            latitude = 41.82592, elevation = 100, slope=0, aspect=0), "spwb_day")